refactor(vscode): move `peckages/` shared code into vscode extension since it's only used by vscode brwoser extension
