+++ editors/vscode/src/browser/shared/document-selector.ts
