+++ editors/vscode/src/browser/shared/initialization-options.ts
