+++ editors/vscode/src/browser/shared/lsp-protocol.ts
