+++ editors/vscode/src/browser/shared/types.ts
