+++ editors/vscode/tsconfig.json
@@ -10,7 +10,6 @@
     "skipLibCheck": true
   },
   "include": [
-    "../../packages/vide-extension-shared/src/**/*.ts",
     "scripts/**/*.ts",
     "src/**/*.ts",
     "test/**/*.ts"
